requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
openai==1.14.3
requests==2.31.0
supabase==2.3.0
Flask-Cors==3.0.10
Pillow==10.3.0 